:bug: Fix crash caused by failure to load file preview icon file
