+++ composeApp/src/commonMain/kotlin/com/clipevery/ui/base/ClipProduceState.kt
@@ -1,0 +1,38 @@
+package com.clipevery.ui.base
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.ProduceStateScope
+import androidx.compose.runtime.State
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import kotlinx.coroutines.suspendCancellableCoroutine
+import kotlin.coroutines.CoroutineContext
+
+class ClipProduceStateScopeImpl<T>(
+    state: MutableState<T>,
+    override val coroutineContext: CoroutineContext,
+) : ProduceStateScope<T>, MutableState<T> by state {
+
+    override suspend fun awaitDispose(onDispose: () -> Unit): Nothing {
+        try {
+            suspendCancellableCoroutine<Nothing> { }
+        } finally {
+            onDispose()
+        }
+    }
+}
+
+@Composable
+fun <T> clipProduceState(
+    initialValue: T,
+    key1: Any?,
+    producer: suspend ProduceStateScope<T>.() -> Unit,
+): State<T> {
+    val result = remember(key1) { mutableStateOf(initialValue) }
+    LaunchedEffect(key1) {
+        ClipProduceStateScopeImpl(result, coroutineContext).producer()
+    }
+    return result
+}
